Fix(Provider): Revert move of provider project custom function logic from web to provider sub-project
