+++ alert-common/src/main/java/com/synopsys/integration/alert/common/descriptor/config/field/endpoint/table/model/ProviderProjectSelectOption.java
@@ -1,5 +1,5 @@
 /**
- * provider
+ * alert-common
  *
  * Copyright (c) 2020 Synopsys, Inc.
  *
@@ -20,15 +20,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.provider.blackduck.web.project;
+package com.synopsys.integration.alert.common.descriptor.config.field.endpoint.table.model;
 
 import com.synopsys.integration.alert.common.rest.model.AlertSerializableModel;
 
-public class BlackDuckProjectSelectOption extends AlertSerializableModel {
+public class ProviderProjectSelectOption extends AlertSerializableModel {
     private final String name;
     private final String description;
 
-    public BlackDuckProjectSelectOption(String name, String description) {
+    public ProviderProjectSelectOption(String name, String description) {
         this.name = name;
         this.description = description;
     }
