+++ web/src/main/java/com/synopsys/integration/alert/web/api/provider/project/ProviderProjectFunctionController.java
@@ -1,5 +1,5 @@
 /**
- * provider
+ * web
  *
  * Copyright (c) 2020 Synopsys, Inc.
  *
@@ -20,22 +20,23 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.provider.blackduck.web.project;
+package com.synopsys.integration.alert.web.api.provider.project;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.synopsys.integration.alert.common.descriptor.config.field.endpoint.table.model.ProviderProjectOptions;
 import com.synopsys.integration.alert.common.descriptor.config.ui.ProviderDistributionUIConfig;
 import com.synopsys.integration.alert.common.rest.api.AbstractFunctionController;
 
 @RestController
-@RequestMapping(ProjectFilterFunctionController.CHANNEL_CONFIGURED_PROJECT_FUNCTION_URL)
-public class ProjectFilterFunctionController extends AbstractFunctionController<BlackDuckProjectOptions> {
+@RequestMapping(ProviderProjectFunctionController.CHANNEL_CONFIGURED_PROJECT_FUNCTION_URL)
+public class ProviderProjectFunctionController extends AbstractFunctionController<ProviderProjectOptions> {
     public static final String CHANNEL_CONFIGURED_PROJECT_FUNCTION_URL = AbstractFunctionController.API_FUNCTION_URL + "/" + ProviderDistributionUIConfig.KEY_CONFIGURED_PROJECT;
 
     @Autowired
-    public ProjectFilterFunctionController(ProjectFilterCustomFunctionAction functionAction) {
+    public ProviderProjectFunctionController(ProviderProjectCustomFunctionAction functionAction) {
         super(functionAction);
     }
